requirements.txt
Django==3.1.5
asgiref==3.3.1
psycopg2-binary
pytz==2020.5
sqlparse==0.4.1
django-cors-headers
djangorestframework
